workers/resume-parser-worker/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
textract==1.6.5
pdfminer.six==20221105
docx2txt==0.8
ebooklib==0.18
beautifulsoup4==4.12.2
lxml==4.9.3

# NLP and AI dependencies
spacy==3.7.2
nltk==3.8.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Data processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
dateutil==2.8.2
regex==2023.10.3

# Job queue and messaging
celery==5.3.4
redis==5.0.1
nats-py==2.7.2

# HTTP client
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Logging and monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Environment specific
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Additional ML libraries
xgboost==2.0.0
lightgbm==4.1.0
catboost==1.2.2

# Computer vision (for scanned resumes)
opencv-python==4.8.1.78
pytesseract==0.3.10
